lab5/pyproject.toml
[project]
name = "lab5"
version = "0.1.0"
description = ""
authors = [
    {name = "romashorodok",email = "108732029+romashorodok@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "kafka-python (>=2.1.5,<3.0.0)"
]

[tool.poetry]
packages = [
    {include = "consumer"},
    {include = "producer"},
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
